This assignment requires that you write a program in C++ to list the contents of a TAR file. The format of TAR files can be found at http://en.wikipedia.org/wiki/Tar_file_format.
Here is a sample run of your program, assuming the name of your program is
mytar.

$ ./mytar /student/test.tar
total 8 files
-rw-r--r-- root root 512 b9134001/512byte
drwxr-xr-x root wheel 0 b9134001/hw1/
-rw-r--r-- root wheel 256 b9134001/hw1/Makefile
-rw-r--r-- root wheel 551 b9134001/hw1/tarfile.h
-rw-r--r-- root wheel 318 b9134001/hw1/main.cpp
-rw-r--r-- root wheel 148 b9134001/hw1/main.h
-rw-r--r-- root wheel 1662 b9134001/hw1/tarfile.cpp
drwx------ b9134001 student 0 b9134001/hw3/
$ ./mytar mytar
Not a ustar file.
File opening error.
$

The output of your program has to be similar to that of the tar command.
Also, your program has to check to see if the file type is “USTAR format”
as described in the document. The test data test.tar can be found in the
directory /student. Note that this is also the file to be used by your TA
to test your program during the demo. You can use the structure given
below to read the file header.